Fix visibility of header files, add module map
Update the podspec such that the public/private/project attributes of the various header files match the actual CareKit.xcodeproj.

Add a module map that refers to CareKit.h as the true umbrella header for the framework so that any modifications to that header are reflected in the CocoaPod as well.