Synchronize Lazy Instantiation of Schedule Calendar
The `UTC_calendar` property on `OCKCareSchedule` is initialized
and configured lazily on first access inside a getter. This operation
is not atomic. Because the framework may call schedule methods on
different threads (for example, some operations are on
the OCKCarePlan serial), this can create a race condition.

To resolve this issue, we synchronizing the lazy initialization on
the instance itself. This resolved crashes we'd been seeing related
to this property.