Resolve locale bug
This resolves a bug in which programmatically setting the displayed date of the week calendar view could cause an infinite recursion in some locales.

The underlying issue was the way in which the beginning of the week was being calculated. The solution was to use a different method for determining the week in which a date falls.