CareKit 2.0 Updates (#313)
We’re happy to introduce lots of new and exciting changes in this commit! As always, feel free to reach out with questions or comments regarding the new changes or the framework as a whole. 

Note that this update brings changes to the schema and is not backwards compatible. Therefore, you must delete and re-install your current app that uses CareKitStore. We will not be providing migrations during this beta period.

Below are some of the major changes:

CareKitUI:
- Added a new simple contact card style.
- Support for Dark Mode.
- `OCKButtonLogTaskView` now has a collection view of log buttons, providing the freedom to add additional log buttons.
- Custom style can now be injected into views. See the `OCKStylable` protocol.

CareKitStore:
- Renamed database properties such as `deletedAt` (to `deletedDate`), `updatedAt` (to `updatedDate`), `effectiveAt` (to `effectiveDate`), `createdAt` (to `createdDate`), and `externalID` (to `remoteID`).
- Added new properties to `OCKPatient` such as sex, birthday, allergies.
- Added ability to fetch and query entities by remote ID, parent entity ID, tag, and group Identifier.
- Added semantic versioning for the database version.
- Removed `versionID` alias.

CareKit:
- Error message are now printing correctly.

Across the Framework:
- Added ability to create custom synchronized cards. For example, to create a custom synchronized task card, see `OCKTaskDisplayer`, `OCKTaskDisplayable`, and `OCKTaskViewController`.
- Bug fixes.