Move store from scene delegate into app delegate
We do not want to create or keep the store in the scene delegate. If the app supports multiple windows (new in iOS 13) a new scene delegate will be instantiated for every window, which means a new store would be created too. We want to share the store across windows, so it should go in the app delegate. This updates the sample app to reflect best practices.