Fix crash caused by updating a task without changing its effective date
These changes modify the behavior of versioned tasks to

1. Avoid a hard crash if a later version of a task has an earlier `effectiveDate` than a previous version.

2. Assure that outcomes can't be added to "shadowed" areas of past versions of tasks.

Consider the following case.

```
|--------- Query Interval ------------|
     Task V1 ---------------------------->
                                  V2 ------------>
                          V3------------------->
```

When querying for events in the given interval, we start from version three and gather all outcomes. We then move on to version 2 and gather all outcomes in the region before V3 begins. A crash would occur in this case because there the time between the start of V2 and V3 is negative.

**This was fixed skipping over versions that are fully shadowed**. Unit tests have been added to prevent regression.

```
|--------- Query Interval ------------|
     Task V1 ---------------------------->
                                  V2 -x------x-->
                          V3------------------->
```

Another possibly unexpected result could occur if there were outcomes saved to V2 when V3 was created. Before the creation of V3, queries would return outcomes from V2, but after the creation of V3, those outcomes would no longer be returned because V2 is shadowed by V3. 

**This was fixed by adding checks that make the transaction fail if updating a task or outcome would result in "shadowed" outcomes.** Unit tests have also been added to ensure the expected behavior in these cases as well.