Fix query by tag bug
This commit fixes a problem in which queries against tags would work on .inMemory stores but not on .onDisk stores.

The underlying problem was that CoreData serializes arrays of tags to binary when they are saved to disk, and predicates do no evaluate properly against binary blobs. In memory stores do not serialize anything, so predicates do work as one would expect.

The solution was to perform filtering against tags at the CareKitStore level instead of the CoreData level. This solution is sub-optimal from a performance perspective, but will likely never manifest as a problem given the number of tasks in a typical CareKit app.

We considered addressing the root problem by changing the tags from an array of strings into a many-to-many relationship with its own table. However, this would require a complicated database migration and doesn't seem worth the effort or the risk. In the future, we may consider making this change if it can be batched with other changes into a single migration.