Address version look up logic error
This commit addresses a logic error in the store's recursive event fetch method. When fetching events from old versions of a task, the query start date was being set to a date that was later than the query applied at the beginning of the recursion. This resulted in invalid date ranges when recursing back more than one task version.