Resolve task version chain bug (#374)
* Ensure task view controllers update when their watched task changes

* Address version look up logic error

This commit addresses a logic error in the store's recursive event fetch method. When fetching events from old versions of a task, the query start date was being set to a date that was later than the query applied at the beginning of the recursion. This resulted in invalid date ranges when recursing back more than one task version.

* Fix single event query bug 

This commit addresses a problem in which fetchEvent(task:occurrence:) would fail if the outcome for the event was nil.

The solution was to use the fetchOutcomes (plural) method instead of fetchOutcome (singular) because the former succeeds with an empty array when no outcomes exist.