Fix a bug where all day schedule elements returned the wrong events (#538)
This fixes a bug that would occur when generating events for an all day schedule element. If generating events for a date range that started several days after the schedule element's start date, and if the schedule element were an all day event, then events would be incorrectly returned starting from the the element's start date up to the query's start date.